# diag/api/swagger.yml
swagger: '2.0'
info:
  description: Diagnostics api
  version: 0.0.0
  title: HW Diag
basePath: /api
paths:
...
  /leds:
    get:
      tags:
        - leds
      summary: List all leds
      description: Returns all leds defined in /sys/class/leds
      operationId: getLeds
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              type: string
            example:
              - led0
              - led1
              - ...
        '501':
          description: Error opening the directory
  /leds/{name}/{color}:
    put:
      tags:
        - leds
      summary: Set the one led's color
      description: Set the intensity of a led
      operationId: updateLedColor
      parameters:
        - name: name
          in: path
          description: Name of the led
          required: true
          type: string
        - name: color
          in: path
          description: intensity between 1 and 100
          required: true
          type: integer
          minimum: 0
          maximum: 100
      responses:
        '200':
          description: Success
        '501':
          description: Error accessing led or parsing parameter
  /leds/all/{color}:
    post:
      tags:
        - leds
      summary: Set multiple leds' color
      description: ''
      operationId: setLedsColor
      parameters:
        - name: color
          in: path
          description: Dash separated intensity between 1 and 100; 69-69-69
          required: true
          type: string
        - in: body
          name: body
          schema:
            type: object
            properties:
              names:
                type: array
                items:
                  type: string
                description: List of led names without color postfix
                example:
                  - led0
                  - led1
              separator:
                type: string
                description: Separator character(s) between led name and color postfix
                example: _
              rString:
                type: string
                description: Red color postfix
                example: r
              gString:
                type: string
                description: Green color postfix
                example: g
              bString:
                type: string
                description: Blue color postfix
                example: b
      responses:
        '200':
          description: Success
        '501':
          description: Error accessing led or parsing parameter
...